add custom test environment and split up tests
